requirements.txt
torch
torchvision
facenet-pytorch
opencv-python-headless
numpy
Pillow
requests
streamlit
